Generalize Logging.hs
