+++ src/Restyled/Development/Seeds.hs
@@ -16,7 +16,6 @@ import qualified Data.List.NonEmpty as NE
 import Data.Monoid (First)
 import Data.Time.Clock.POSIX (utcTimeToPOSIXSeconds)
 import Restyled.Env
-import Restyled.Logging
 import Restyled.Marketplace
 import Restyled.Models
 import Restyled.PrivateRepoAllowance
@@ -88,7 +87,8 @@ runApp f = do
         app = "seed-db"
 
         runLogging :: LoggingT IO a -> IO a
-        runLogging = runAppLoggingT logger . withThreadContext ["app" .= app]
+        runLogging =
+            runLoggerLoggingT logger . withThreadContext ["app" .= app]
 
     appSqlPool <- runLogging $ do
         createConnectionPool appDatabaseConf appStatementTimeout
