+++ src/Restyled/SyncMarketplace.hs
@@ -11,7 +11,6 @@ import qualified Data.Vector as V
 import qualified GitHub.Endpoints.MarketplaceListing.Plans as GH
 import qualified GitHub.Endpoints.MarketplaceListing.Plans.Accounts as GH
 import Restyled.Env
-import Restyled.Logging
 import Restyled.Models
 import Restyled.PrivateRepoAllowance
 import Restyled.Tracing
@@ -78,7 +77,8 @@ runApp f = do
         app = "sync-marketplace"
 
         runLogging :: LoggingT IO a -> IO a
-        runLogging = runAppLoggingT logger . withThreadContext ["app" .= app]
+        runLogging =
+            runLoggerLoggingT logger . withThreadContext ["app" .= app]
 
     appSqlPool <- runLogging
         $ createConnectionPool appDatabaseConf appStatementTimeout
